requirements.txt
boto3
botocore
certifi
chardet
dj-database-url
Django
django-crispy-forms
django-heroku
django-storages
docutils
gunicorn
idna
jmespath
Pillow
psycopg2
python-dateutil
pytz
requests
s3transfer
six
urllib3
whitenoise
djangorestframework
